pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nethang"
version = "0.1.1"
description = "A web-based tool for simulating network quality"
readme = "README.md"
authors = [
    {name = "Hang Yin", email = "stephen.yin.h@gmail.com"}
]
license = "MIT"
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Operating System :: POSIX :: Linux",
    "Operating System :: POSIX :: Linux :: Debian",
    "Operating System :: POSIX :: Linux :: Ubuntu",
]
dependencies = [
    "Flask>=2.0.0",
    "flask-socketio>=5.0.0",
    "python-dotenv>=0.19.0",
    "Werkzeug>=2.0.0",
    "PyYAML>=6.0.0",
    "tomli>=2.0.0",
]

[project.scripts]
nethang = "run:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "mypy>=1.0.0",
    "types-PyYAML>=6.0.0",
    "black>=22.0.0",
    "flake8>=4.0.0",
    "isort>=5.0.0",
]

[project.urls]
Homepage = "https://github.com/stephenyin/NetHang"
Repository = "https://github.com/stephenyin/NetHang.git"

[tool.setuptools]
packages = ["nethang"]
package-dir = {"" = "."}

[tool.black]
line-length = 88
target-version = ["py38"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "--cov=nethang --cov-report=term-missing"
